Updated UserDAO, CourseDAO, and AnnouncementDao to include @insert, @update and @delete for testing
